Add Resource Parameter Overriding Support
Added support for hash based resource assignment.
-- Ex: File["name"] { key => val }

Separated scoping related functions, so scopes can be checked for
resources as well from within the Catalog module.

Added support for resource parameter overriding from child classes
-- Ex: class a { file { "foo": key => val1 } }
--     class b inherits a { File["foo"] { key => val2 } }

Fixed a bug in the test suite which caused files tagged with Expect Fail
and Unsupported together to be marked OK if both failed. With this being
an Unsupported feature this is incorrect. A custom message was also
added for those cases.